Backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.10.0
asttokens==3.0.0
attrs==25.4.0
beautifulsoup4==4.14.2
bs4==0.0.2
certifi==2025.11.12
cffi==1.17.1
charset-normalizer==3.4.4
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cryptography==45.0.5
curl_cffi==0.14.0
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
dj-rest-auth==7.0.1
Django==5.2.8
django-allauth==65.13.1
django-cors-headers==4.9.0
django-environ==0.12.0
django-extensions==4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
executing==2.2.1
faiss-cpu==1.13.2
finance-datareader==0.9.100
fonttools==4.60.1
frozendict==2.4.7
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipykernel==6.30.1
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.12.0
joblib==1.5.3
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
lxml==6.0.2
matplotlib==3.10.7
matplotlib-inline==0.2.1
multitasking==0.0.12
narwhals==2.14.0
nest-asyncio==1.6.0
numpy==2.3.3
oauthlib==3.3.1
openai==2.14.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
parso==0.8.5
peewee==3.18.3
pillow==12.0.0
platformdirs==4.4.0
plotly==6.5.0
prompt_toolkit==3.0.52
protobuf==6.33.2
psutil==7.1.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
pyzmq==27.1.0
requests==2.32.5
requests-file==3.0.1
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
selenium==4.38.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
sqlparse==0.5.3
stack-data==0.6.3
threadpoolctl==3.6.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.32.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Wikipedia-API==0.8.1
wsproto==1.3.2
yfinance==0.2.66
